refactor(shared): reuse node list parsers across cli and tools
